Fix signedness warnings

+++ src/test/objectstore/ObjectStoreImitator.cc
@@ -13,7 +13,7 @@
 #define OBJECT_MAX_SIZE 0xffffffff // 32 bits
 
 void ObjectStoreImitator::init_alloc(const std::string &alloc_type,
-                                     int64_t size) {
+                                     uint64_t size) {
   alloc.reset(Allocator::create(cct, alloc_type, size, min_alloc_size));
   alloc->init_add_free(0, size);
   ceph_assert(alloc->get_free() == size);
@@ -747,7 +747,7 @@ int ObjectStoreImitator::_collection_list(
 
   do {
     ls->push_back((it++)->first);
-    if (ls->size() >= max) {
+    if (ls->size() >= (unsigned)max) {
       *next = it->first;
       return 0;
     }
